+++ Deplora.WPF/ViewModels/MainWindowViewModel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Deplora.WPF.Commands;
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Windows.Input;
@@ -12,8 +13,20 @@ namespace Deplora.WPF.ViewModels
 
         public MainWindowViewModel()
         {
-            this.ShowAppSettings = new Commands.ShowAppSettingsCommand();
-            this.ShowDeployConfigurations = new Commands.ShowManageDeployConfigurationsCommand();
+            this.ShowAppSettings = new RelayCommand(ShowAppSettingsWindow);
+            this.ShowDeployConfigurations = new RelayCommand(ShowDeployConfigurationsWindow);
+        }
+
+        private void ShowAppSettingsWindow()
+        {
+            var appSettings = new ApplicationSettings();
+            appSettings.Show();
+        }
+
+        private void ShowDeployConfigurationsWindow()
+        {
+            var deployConfigurationsWindow = new DeployConfigurationList();
+            deployConfigurationsWindow.Show();
         }
     }
 }
